requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
Django==3.2.7
idna==3.2
image==1.5.33
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
pytz==2021.3
pyzbar==0.1.8
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
